Move makefile instructions into tox
The makefile was usable only on linux and with this change we can
run them on other OSes. The instructions from the makefile were moved
in tox, one of the de facto ways of running tests in Python.
We're also separating the generation of the coverage with the
upload of the coverage.